requirements.txt
# --- Core Backend (FastAPI) ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic[email]
pydantic-settings

# --- Database (PostgreSQL & ORM) ---
sqlalchemy==2.0.28
asyncpg==0.29.0
black
alembic==1.13.1
sqladmin==0.18.0
psycopg2-binary==2.9.9

# --- Authentication (JWT) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# --- Real-time (WebSockets) ---
websockets==12.0

# --- Task Queue (Celery & Background tasks) ---
celery==5.3.6
redis==5.0.3

# --- Додатково для Celery & Monitoring ---
flower==2.0.1        # Моніторинг тасок у браузері
eventlet==0.35.1     # Для стабільної роботи воркера (особливо на Windows)

# --- Додатково для SQLAlchemy ---
greenlet==3.0.3      # Потрібен для асинхронних проксі-об'єктів

# --- Робота з часом та часовими поясами ---
pytz==2024.1         # Стандарт для роботи з Timezones

# --- Telegram Bot (Moderation) ---
aiogram

# --- Rate limiting ---
slowapi==0.1.9

# --- Testing ---
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# --- File utilities ---
aiofiles==23.2.1

# --- Utilities & Tools ---
python-dotenv==1.0.1
httpx==0.27.0
pillow==10.2.0